[home_work_1] add Base Fragment, fix tv_shows_fragment.xml, TvShowsFragment
