+++ app/src/main/java/ru/androidschool/intensiv/ui/tvshows/TvShowsFragment.kt
@@ -4,23 +4,17 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
 import androidx.navigation.navOptions
 import com.xwray.groupie.GroupAdapter
 import com.xwray.groupie.GroupieViewHolder
 import ru.androidschool.intensiv.R
+import ru.androidschool.intensiv.base.BaseFragment
 import ru.androidschool.intensiv.data.MockRepository
-import ru.androidschool.intensiv.data.Movie
 import ru.androidschool.intensiv.data.ShowDest
-import ru.androidschool.intensiv.databinding.FeedFragmentBinding
 import ru.androidschool.intensiv.databinding.TvShowsFragmentBinding
-import ru.androidschool.intensiv.ui.feed.FeedFragment.Companion.KEY_TITLE
-import ru.androidschool.intensiv.ui.feed.MovieItem
 
-class TvShowsFragment : Fragment(R.layout.tv_shows_fragment) {
-    private var _binding: TvShowsFragmentBinding? = null
-    private val binding get() = _binding!!
+class TvShowsFragment : BaseFragment<TvShowsFragmentBinding>() {
 
     private val adapter by lazy {
         GroupAdapter<GroupieViewHolder>()
@@ -35,13 +29,11 @@ class TvShowsFragment : Fragment(R.layout.tv_shows_fragment) {
         }
     }
 
-    override fun onCreateView(
+    override fun createViewBinding(
         inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        _binding = TvShowsFragmentBinding.inflate(inflater, container, false)
-        return binding.root
+        container: ViewGroup?
+    ): TvShowsFragmentBinding {
+        return TvShowsFragmentBinding.inflate(inflater, container, false)
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
@@ -64,11 +56,6 @@ class TvShowsFragment : Fragment(R.layout.tv_shows_fragment) {
         val bundle = Bundle()
         bundle.putString(KEY_TITLE, showDest.title)
         findNavController().navigate(R.id.movie_details_fragment, bundle, options)
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
     }
 
     companion object {
